# webui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Swedish Sentiment API",
    "description": "REST API for Swedish sentiment analysis, ASR transcription, call-center conversation analysis, and batch processing.",
    "version": "0.5.1"
  },
  "paths": {
...
    "/transcription/jobs": {
      "get": {
        "tags": [
          "Transcription"
        ],
        "summary": "List Transcription Jobs",
        "description": "List recent transcription jobs (newest first).",
        "operationId": "list_transcription_jobs_transcription_jobs_get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 20,
              "title": "Limit"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeJobListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transcription/jobs/{job_id}": {
      "get": {
        "tags": [
          "Transcription"
        ],
        "summary": "Get Transcription Job",
        "description": "Return status snapshot for a transcription job.",
        "operationId": "get_transcription_job_transcription_jobs__job_id__get",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeJobStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/transcription/jobs/{job_id}/cancel": {
      "post": {
        "tags": [
          "Transcription"
        ],
        "summary": "Cancel Transcription Job",
        "description": "Request cancellation of a running transcription job.",
        "operationId": "cancel_transcription_job_transcription_jobs__job_id__cancel_post",
        "security": [
          {
            "APIKeyHeader": []
          }
        ],
        "parameters": [
          {
            "name": "job_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Job Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeJobCancelResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/upload": {
      "post": {
        "tags": [
          "Transcription"
        ],
        "summary": "Upload Audio File",
        "description": "Upload an audio file to the server for transcription.\n\nThe file is saved under API_MEDIA_ROOT/uploads/ with a unique filename.\nReturns the validated server-side path that can be used with POST /transcribe.\n\nSupports common audio formats: wav, mp3, m4a, flac, ogg, webm.\nMaximum file size: API_MAX_UPLOAD_SIZE_MB (default 200 MB).\nOld files are cleaned up after API_UPLOAD_RETENTION_DAYS (default 7 days).",
        "operationId": "upload_audio_file_upload_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_audio_file_upload_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/transcribe": {
      "post": {
        "tags": [
          "Transcription"
        ],
        "summary": "Transcribe",
        "description": "Transcribe an audio file using ASR.",
        "operationId": "transcribe_transcribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
    "/batch_transcribe": {
      "post": {
        "tags": [
          "Transcription"
        ],
        "summary": "Batch Transcribe",
        "description": "Transcribe multiple audio files, optionally in parallel.",
        "operationId": "batch_transcribe_batch_transcribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BatchTranscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BatchTranscribeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "APIKeyHeader": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Body_upload_audio_file_upload_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "description": "Audio file to upload"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_audio_file_upload_post"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TranscribeJobCancelResponse": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "cancelled",
          "timestamp"
        ],
        "title": "TranscribeJobCancelResponse"
      },
      "TranscribeJobListResponse": {
        "properties": {
          "jobs": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Jobs"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "jobs",
          "timestamp"
        ],
        "title": "TranscribeJobListResponse"
      },
      "TranscribeJobStatus": {
        "properties": {
          "job_id": {
            "type": "string",
            "title": "Job Id"
          },
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "cancelled": {
            "type": "boolean",
            "title": "Cancelled",
            "default": false
          },
          "meta": {
            "additionalProperties": true,
            "type": "object",
            "title": "Meta"
          }
        },
        "type": "object",
        "required": [
          "job_id",
          "kind",
          "status",
          "created_at"
        ],
        "title": "TranscribeJobStatus"
      },
      "TranscribeRequest": {
        "properties": {
          "model": {
            "type": "string",
            "title": "Model",
            "default": "kb-whisper-large"
          },
          "backend": {
            "type": "string",
            "title": "Backend",
            "description": "faster | transformers | whisperx (alignment + diarization)",
            "default": "faster"
          },
          "device": {
            "type": "string",
            "title": "Device",
            "default": "auto"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "sv"
          },
          "beam_size": {
            "type": "integer",
            "maximum": 10.0,
            "minimum": 1.0,
            "title": "Beam Size",
            "default": 5
          },
          "vad": {
            "type": "boolean",
            "title": "Vad",
            "default": true
          },
          "chunk_length_s": {
            "type": "integer",
            "maximum": 60.0,
            "minimum": 5.0,
            "title": "Chunk Length S",
            "default": 30
          },
          "revision": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Revision",
            "description": "KB-Whisper revision: standard|strict|subtitle"
          },
          "diarize": {
            "type": "boolean",
            "title": "Diarize",
            "description": "Run speaker diarization",
            "default": false
          },
          "num_speakers": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Num Speakers",
            "description": "Expected number of speakers (None=auto)"
          },
          "hotwords": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hotwords",
            "description": "Domain-specific words to boost (callcenter terms etc.)"
          },
          "initial_prompt": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Initial Prompt",
            "description": "Conditioning prompt for ASR decoder"
          },
          "provider": {
            "type": "string",
            "enum": [
              "local",
              "cloud"
            ],
            "title": "Provider",
            "description": "ASR provider: local (default) | cloud (Deepgram, opt-in)",
            "default": "local"
          },
          "cloud_fallback_local": {
            "type": "boolean",
            "title": "Cloud Fallback Local",
            "description": "Fall back to local ASR if cloud fails (default off)",
            "default": false
          },
          "audio_path": {
            "type": "string",
            "title": "Audio Path",
            "description": "Path to audio file accessible by the server"
          },
          "original_filename": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Filename",
            "description": "Original upload filename; used for customer identification when a customer registry is active"
          },
          "word_timestamps": {
            "type": "boolean",
            "title": "Word Timestamps",
            "default": true
          },
          "preprocess": {
            "type": "boolean",
            "title": "Preprocess",
            "description": "Enable audio preprocessing before ASR",
            "default": false
          },
          "preprocess_mode": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Preprocess Mode",
            "description": "Preprocess mode: off | basic | callcenter (v2 bandpass + tuned VAD). Overrides legacy boolean when set."
          },
          "run_partial_analysis": {
            "type": "boolean",
            "title": "Run Partial Analysis",
            "description": "After ASR, run incremental partial pipeline on transcript segments (local path)",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "audio_path"
        ],
        "title": "TranscribeRequest"
      },
...
      "UploadResponse": {
        "properties": {
          "audio_path": {
            "type": "string",
            "title": "Audio Path",
            "description": "Validated server-side path to uploaded file"
          },
          "filename": {
            "type": "string",
            "title": "Filename",
            "description": "Original filename from upload"
          },
          "size_bytes": {
            "type": "integer",
            "title": "Size Bytes",
            "description": "File size in bytes"
          },
          "customer": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CustomerRef"
              },
              {
                "type": "null"
              }
            ],
            "description": "Resolved customer context when a customer registry is active"
          },
          "timestamp": {
            "type": "string",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "audio_path",
          "filename",
          "size_bytes",
          "timestamp"
        ],
        "title": "UploadResponse",
        "description": "Response from POST /upload."
      },
...
    },
    "securitySchemes": {
      "APIKeyHeader": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key"
      }
    }
  }
}